requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# LLM & Agents
langchain>=0.1.9
langchain-openai>=0.0.6
langchain-community>=0.0.24
langgraph>=0.0.26
openai>=1.12.0

# Vector Database (ChromaDB - local, easy to start)
chromadb==0.4.22

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# GitHub Integration
PyGithub==2.2.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
structlog==24.1.0

# Code Parsing
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.2

# Frontend
streamlit>=1.30.0
plotly>=5.18.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0

# Docker/Deployment
gunicorn==21.2.0
